Converting Inventory ro SO
