+++ Assets/InventorySystem/Example/Data/Inventories.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 87e408673b08b8e458d5e71c94d4236a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
