+++ Assets/InventorySystem/Example/Data/Inventories/CraftInventory.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3790736a552e0644da5724c0ac7b9d21
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
