+++ Assets/InventorySystem/Example/Data/Inventories/Inventory 2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c040294a07fd2a64387d1d83bcf31b55
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
