+++ Assets/InventorySystem/Example/Data/Inventories/Inventory.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b454f7e4246ea4341bdeb4dc5c9fb7a4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
