+++ Assets/InventorySystem/Example/ExampleScene.unity
@@ -1512,77 +1512,7 @@ MonoBehaviour:
   dropOnCloseCrafting: 0
   dropPos: {x: 0, y: 0, z: 0}
   randomFactor: {x: 1, y: 1, z: 0}
-  inv:
-    slots:
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    slotAmounts: 9
-    id: 0
-    areItemsUsable: 0
-    areItemsDroppable: 1
-    interactiable: -1
-    hasInitializated: 0
+  inv: {fileID: 11400000, guid: 3790736a552e0644da5724c0ac7b9d21, type: 2}
   isCraftInventory: 1
   gridSize: {x: 3, y: 3}
   allowsPatternCrafting: 1
@@ -3919,70 +3849,7 @@ MonoBehaviour:
   dropOnCloseCrafting: 0
   dropPos: {x: 0, y: 0, z: 0}
   randomFactor: {x: 0, y: 0, z: 0}
-  inv:
-    slots:
-    - amount: 1
-      item: {fileID: 11400000, guid: 5e31591f363c6064e82e01bb211d7c50, type: 2}
-      hasItem: 1
-      _durability: 8
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    - amount: 0
-      item: {fileID: 0}
-      hasItem: 0
-      _durability: 0
-      isProductSlot: 0
-      interative: -1
-      whitelist: {fileID: 0}
-    slotAmounts: 8
-    id: 0
-    areItemsUsable: 1
-    areItemsDroppable: 1
-    interactiable: -1
-    hasInitializated: 0
+  inv: {fileID: 11400000, guid: c040294a07fd2a64387d1d83bcf31b55, type: 2}
   isCraftInventory: 0
   gridSize: {x: 0, y: 0}
   allowsPatternCrafting: 0
@@ -4620,14 +4487,7 @@ MonoBehaviour:
   dropOnCloseCrafting: 0
   dropPos: {x: 0, y: 0, z: 0}
   randomFactor: {x: 0, y: 0, z: 0}
-  inv:
-    slots: []
-    slotAmounts: 0
-    id: 0
-    areItemsUsable: 0
-    areItemsDroppable: 0
-    interactiable: 1
-    hasInitializated: 0
+  inv: {fileID: 11400000, guid: b454f7e4246ea4341bdeb4dc5c9fb7a4, type: 2}
   isCraftInventory: 0
   gridSize: {x: 0, y: 0}
   allowsPatternCrafting: 0
@@ -5873,14 +5733,7 @@ MonoBehaviour:
   dropOnCloseCrafting: 0
   dropPos: {x: 0, y: 0, z: 0}
   randomFactor: {x: 0, y: 0, z: 0}
-  inv:
-    slots: []
-    slotAmounts: 0
-    id: 0
-    areItemsUsable: 1
-    areItemsDroppable: 1
-    interactiable: -1
-    hasInitializated: 0
+  inv: {fileID: 11400000, guid: b454f7e4246ea4341bdeb4dc5c9fb7a4, type: 2}
   isCraftInventory: 0
   gridSize: {x: 0, y: 0}
   allowsPatternCrafting: 0
@@ -5915,8 +5768,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 8ac0707fd66611846838e9c9b6a98c25, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  invUI: {fileID: 1140438006}
-  invUI2: {fileID: 882816579}
+  inventory: {fileID: 11400000, guid: b454f7e4246ea4341bdeb4dc5c9fb7a4, type: 2}
   testItem: {fileID: 11400000, guid: 66afe602a9207284f85d380a8f01c58c, type: 2}
   slotAmount: 10
 --- !u!114 &1140438009
