+++ Assets/InventorySystem/Materials/DragItemMat.mat
@@ -75,5 +75,5 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 0.764151, g: 0.37030083, b: 0.2270826, a: 1}
+    - _Color: {r: 0.1490196, g: 0.5592764, b: 0.77254903, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
