+++ Assets/InventorySystem/Scripts/Inventory/InventoryController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b278c471a3efbda41a30ef8f936d20a0
+guid: 6ab16c779dd601842b5d5eddb6cbb7b4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
